Restructure more components/services
