+++ src/client/app/admin/end-user-agreement/admin-list-euas.component.ts
@@ -6,11 +6,13 @@ import * as _ from 'lodash';
 import { DialogRef } from 'angular2-modal';
 import { Modal } from 'angular2-modal/plugins/bootstrap';
 
-import { SortDisplayOption, SortDirection } from 'app/shared/result-utils.class';
-import { AlertService, PagingOptions, TableSortOptions } from 'app/shared';
+import {
+	SortDisplayOption, SortDirection,
+	AlertService, PagingOptions, TableSortOptions
+} from 'app/shared';
 import { AuthenticationService } from '../authentication';
-import { AdminService } from '../admin.service';
 
+import { AdminService } from '../admin.service';
 import { EndUserAgreement } from './eua.class';
 import { EuaService } from './eua.service';
