+++ src/client/app/notifications/notification.class.ts
@@ -1,4 +1,4 @@
-import { Resource } from 'app/resources/resource.class';
+import { Resource } from 'app/resources';
 
 export class Notification {
 	public _id: string;
